[GE] Updates

+++ Engine/Source/Runtime/Events/Public/KeyCodes.h
@@ -2,7 +2,7 @@
 
 namespace Goss
 {
-	using KeyCode = uint16_t;
+	using KeyCode = unsigned int;
 	
 	namespace Key
 	{
